requirements.txt
cfgv==3.4.0
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
distlib==0.4.0
Farama-Notifications==0.0.4
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
gymnasium==1.2.1
# # Editable install with no version control (gymnasium_env==0.0.1)
# -e /mnt/data/Documents/Project_M/Project M/0. Custom Env/Custom Env
identify==2.6.14
imageio==2.37.0
imageio-ffmpeg==0.6.0
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
moviepy==2.2.1
mpmath==1.3.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
pre_commit==4.3.0
proglog==0.1.12
pygame==2.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
virtualenv==20.34.0
